Third cleaning pass
This is shaping up nicely. Hopefully I'll be able to get this out before
classes start. I'd still like to set up a test suite akin to what pass
has, as well as allow the autogenerated completions to support the
completion of filenames from the store. Before that, however, I need to
work on documenting everything, fleshing out the README, and start using
this in place of {go,}pass to ensure it is actually usable day-to-day.

* I decided against allowing the option of wiping the store and have
thus reinstated the `gpg_id` requirement of the `init` subcmd. My
rationale behind this is that 1) pass requires an empty `gpg_id(s)` arg
(not missing, but empty -- ""); and 2) if you were to deinit your store
right this moment, it would delete the root .gpg-id file, throw an error
that you must initialize your password store before using it, and print
out help. Basically, it removes one single file, commits that, and spews
a whole bunch of stuff to stdout. I'd be more inclined to support
reinitialization (with an empty `gpg_id`) than this. But for now,
neither of those are supported.

* Show similar output to when `git commit` is used, showing the first 7
digits of the commit hash, the commit message, followed by a summarized,
shortstat'd diff. Similarity is not yet exposed in `git2-rs`, so it will
always display 100% similarity for now. The rename detection isn't as
sophisticated as `git`, so sometimes it might appear as "deleted file,
added file" when using the `mv` subcommand (that, or I've messed
something up).

* If the user has `commit.gpgsign` set to `true` in their gpg config,
sign commits. Otherwise, don't.

* Drop `rand` as a dependency and use `ring::rand` for random
generation.

* Add `libc` as a dependency so I can get the user's uid and compare
against the specified path passed to `util::set_recursive_permissions`
in order to check if the user owns a path before setting
permissions (would otherwise error if the user didn't own the path).

* Add `psutil` as a dependency so I can kill any previously-running
instances of pass-rs when utilizing the clipboard (to prevent races to
clear the clipboard -- e.g. you clip something, then clip something
else, the first instance of `unclip` would still clear the clipboard
after the timeout, thus clearing the newly clipped contents).

* Fix PASSWORD_STORE_SIGNING_KEY implementation: it is only used to sign
the `.gpg-id` file, not individual entries to the store.

* Allow the generation of a first-time OTP code via the `--generate` (or
`-g`) flag when using `otp::append` or `otp::insert`, to allow for
easier verification of the code. Additionally, display the time left on
that specific code with a fancy little "progress" bar using `format!`
magic (not really).

* Create a `Flags` struct for functions that take more than one boolean
flag because clippy was complaining that 8 is too many args for
`otp::insert`.

* Remove leftover empty dirs when using the `mv` subcmd if there are no
other files in that dir.

* Some refactoring of the `init` subcmd, moving `recrypt_{file,dir}`
into the util module.

* Add licenses for my use of code from `tui-rs`, `treeline`, `copy_dir`,
`grep-cli`, and `terminal_qrcode`.

+++ src/subcmds/cp.rs
@@ -1,32 +1,23 @@
 use std::fs;
-use std::io;
-use std::io::Write;
 use std::os::unix::fs::OpenOptionsExt;
 
 use anyhow::Result;
-use termion::input::TermRead;
 
 use crate::consts::PASSWORD_STORE_UMASK;
 use crate::util;
 use crate::PassrsError;
 
-pub fn cp(force: bool, source: String, dest: String) -> Result<()> {
+pub fn cp(source: String, dest: String, force: bool) -> Result<()> {
     let source_path = util::canonicalize_path(&source)?;
     let is_file = match fs::metadata(&source_path) {
         Ok(meta) => meta.is_file(),
         Err(_) => false,
     };
-
     let dest_path = if is_file {
         util::exact_path(&format!("{}.gpg", dest))?
     } else {
         util::exact_path(&dest)?
     };
-
-    let stdin = io::stdin();
-    let mut stdin = stdin.lock();
-    let stdout = io::stdout();
-    let mut stdout = stdout.lock();
 
     if is_file {
         if !util::path_exists(&source_path)? {
@@ -34,26 +25,20 @@ pub fn cp(force: bool, source: String, dest: String) -> Result<()> {
         }
 
         if !force && util::path_exists(&dest_path)? {
-            write!(
-                stdout,
-                "An entry exists for {}. Overwrite it? [y/N] ",
-                dest_path.display()
-            )?;
-            io::stdout().flush()?;
+            let display = dest_path.display();
+            let prompt = format!("An entry exists for {}. Overwrite it?", display);
 
-            match stdin.read_line()? {
-                Some(reply) if reply.starts_with('y') || reply.starts_with('Y') => {
-                    fs::OpenOptions::new()
-                        .mode(0o666 - (0o666 & *PASSWORD_STORE_UMASK))
-                        .write(true)
-                        .truncate(true)
-                        .open(&dest_path)?;
-                }
-                _ => return Err(PassrsError::UserAbort.into()),
+            if util::prompt_yesno(prompt)? {
+                fs::OpenOptions::new()
+                    .mode(0o666 - (0o666 & *PASSWORD_STORE_UMASK))
+                    .write(true)
+                    .truncate(true)
+                    .open(&dest_path)?;
+            } else {
+                return Err(PassrsError::UserAbort.into());
             }
         }
 
-        // Copy file from source_path to dest_path
         util::copy(&source_path, &dest_path, None)?;
         util::commit(format!("Copy {} to {}", source, dest))?;
     } else {
@@ -62,22 +47,16 @@ pub fn cp(force: bool, source: String, dest: String) -> Result<()> {
         }
 
         if !force && util::path_exists(&dest_path)? {
-            write!(
-                stdout,
-                "An entry exists for {}. Overwrite it? [y/N] ",
-                dest_path.display()
-            )?;
-            io::stdout().flush()?;
+            let display = dest_path.display();
+            let prompt = format!("An entry exists for {}. Overwrite it?", display);
 
-            match stdin.read_line()? {
-                Some(reply) if reply.starts_with('y') || reply.starts_with('Y') => {
-                    fs::remove_dir_all(&dest_path)?;
-                }
-                _ => return Err(PassrsError::UserAbort.into()),
+            if util::prompt_yesno(prompt)? {
+                fs::remove_dir_all(&dest_path)?;
+            } else {
+                return Err(PassrsError::UserAbort.into());
             }
         }
 
-        // Recursively copy folder from source_path to dest_path
         util::copy(&source_path, &dest_path, None)?;
         util::commit(format!("Copy {} to {}", source, dest))?;
     }
